refactor: move connect/disconnect into server
